Improve the TF of the skeletons
Now publishing the TF is optional (dynamic reconfigure), as it may get expensive in CPU & network traffic.

Also all the transforms are published in one single publication call from the same transform broadcaster. This lowers the publication from 500hz to 50Hz... and that's with one skeleton, there can be up to 4 I think.

Also now the tf frames have a number appended to them to keep track of which is which (otherwise two users would have the same frames and they would jump like crazy).